+++ R/languages.R
@@ -4,4 +4,14 @@
 #' @export
 mdt_languages <- function() {
   c("EN", "DE", "DE_F", "FR", "RU")
+  c(
+    English = "en",
+    French = "fr",
+    `German (informal)` = "de",
+    `German (formal)` = "de_f",
+    Italian = "it",
+    Russian = "ru",
+    Spanish = "es",
+    Latvian = "lv"
+  )
 }