+++ R/standalone.R
@@ -9,9 +9,8 @@
 #' at the bottom of the screen so that online participants can ask for help.
 #' @param languages (Character vector)
 #' Determines the languages available to participants.
-#' Possible languages include English (\code{"EN"}), French (\code{"FR"}), French (\code{"IT"}),
-#' and German (informal: \code{"DE"}, formal: \code{"DE_F"}).
-#' The first language is selected by default
+#' The first language is selected by default.
+#' Defaults to the value returned by \code{\link{mdt_languages}()}.
 #' @param dict The psychTestR dictionary used for internationalisation.
 #' @param ... Further arguments to be passed to \code{\link{mdt}()}.
 #' @export